Refactor, lint fixes  and extract tests
