+++ soluciones/alexhermida/reto-3-python/README.md
@@ -1,0 +1,24 @@
+# Solución reto 3 en Python
+
+Requisitos:
+    * Python >= 3.6
+
+
+Instalación:
+
+```
+$python -m venv .venv && .venv/bin/activate && pip install -r requirements.txt
+```
+
+
+Ejecutar código:
+
+```bash
+$python main.py
+```
+
+Ejecutar tests:
+
+```bash
+$pytest
+```
