+++ soluciones/alexhermida/reto-3-python/requirements.txt
@@ -1,1 +1,2 @@
+mypy
 pytest
